Fix : table component화
2024-10-22 antd Table component화
