+++ src/components/tables/index.tsx
@@ -1,5 +1,4 @@
-import { Space, Table, Tag } from 'antd';
-import type { TableProps } from 'antd';
+import { Table } from 'antd';
 import { useRouter } from 'next/router';
 
 interface DataType {
@@ -10,93 +9,18 @@ interface DataType {
   address: string;
   tags: string[];
 }
+interface TableColumnProps {
+  columns?: any;
+  dataSource?: any;
+}
 
-const Tables = () => {
+const Tables = ({ columns, dataSource }: TableColumnProps) => {
   const router = useRouter();
-  const columns: TableProps<DataType>['columns'] = [
-    {
-      title: '성함',
-      dataIndex: 'name',
-      key: 'name',
-    },
-    {
-      title: '아이디',
-      dataIndex: 'email',
-      key: 'email',
-    },
-    {
-      title: '전화번호',
-      dataIndex: 'phone',
-      key: 'phone',
-    },
-    {
-      title: '등급',
-      dataIndex: 'ranking',
-      key: 'ranking',
-    },
-    {
-      title: '권한',
-      key: 'tags',
-      dataIndex: 'tags',
-      render: (_, { tags }) => (
-        <>
-          {tags.map((tag) => {
-            let color = tag.length > 5 ? 'geekblue' : 'green';
-            if (tag === 'loser') {
-              color = 'volcano';
-            }
-            return (
-              <Tag color={color} key={tag}>
-                {tag.toUpperCase()}
-              </Tag>
-            );
-          })}
-        </>
-      ),
-    },
-    {
-      title: 'Action',
-      key: 'action',
-      render: (_, record) => (
-        <Space size="middle">
-          <a>호스트 전환</a>
-          <a>탈퇴</a>
-        </Space>
-      ),
-    },
-  ];
-
-  const data: DataType[] = [
-    {
-      key: '1',
-      name: 'John Brown',
-      email: 'test1@gmail.com',
-      phone: '010-1234-5678',
-      address: 'New York No. 1 Lake Park',
-      tags: ['nice', 'developer'],
-    },
-    {
-      key: '2',
-      name: 'Jim Green',
-      email: 'test2@gmail.com',
-      phone: '010-4567-8910',
-      address: 'London No. 1 Lake Park',
-      tags: ['loser'],
-    },
-    {
-      key: '3',
-      name: 'Joe Black',
-      email: 'test3@gmail.com',
-      phone: '010-9876-5432',
-      address: 'Sydney No. 1 Lake Park',
-      tags: ['cool', 'teacher'],
-    },
-  ];
 
   return (
     <Table<DataType>
       columns={columns}
-      dataSource={data}
+      dataSource={dataSource}
       onRow={(record, rowIndex) => {
         return {
           onClick: (e) => {
